refactor: move munkres.options.ts to types/munkresOptions.ts
Relocate the MunkresOptions interface into the types/ directory and
rename it to camelCase to match the other type modules. Update the
import in munkres.ts, the re-export in index.ts, the dispatcher comment,
and the README links. No behavior change.
